Client Codebase Refactor: Consistent Naming, Named Exports, and Auth Logic Cleanup for Maintainability

+++ apps/server/src/modules/auth/features/verify-email/verify-email.service.ts
@@ -1,4 +1,5 @@
 import type { Prisma } from "@prisma/client";
+import { inject, injectable } from "inversify";
 
 import { BadRequestError, ForbiddenError } from "@/modules/shared/api-errors";
 import { ERROR_MESSAGES } from "@/modules/shared/constants";
@@ -14,37 +15,24 @@ import {
 } from "@/modules/auth/constants";
 import type { IJwtService } from "@/modules/auth/utils";
 
+import { TYPES } from "@/config/inversify/types";
+
 import type {
   IVerifyEmailRepository,
   IVerifyEmailService,
-  IVerifyEmailServiceOptions,
 } from "./verify-email.types";
 
+@injectable()
 export class VerifyEmailService implements IVerifyEmailService {
-  private static _instance: VerifyEmailService;
-
-  private constructor(
+  constructor(
+    @inject(TYPES.VerifyEmailRepository)
     private readonly _verifyEmailRepository: IVerifyEmailRepository,
+    @inject(TYPES.PrismaService)
     private readonly _prismaService: IPrismaService,
-    private readonly _userService: IUserService,
-    private readonly _jwtService: IJwtService,
-    private readonly _loggerService: ILoggerService
+    @inject(TYPES.UserService) private readonly _userService: IUserService,
+    @inject(TYPES.JwtService) private readonly _jwtService: IJwtService,
+    @inject(TYPES.LoggerService) private readonly _loggerService: ILoggerService
   ) {}
-
-  public static getInstance(
-    options: IVerifyEmailServiceOptions
-  ): VerifyEmailService {
-    if (!this._instance) {
-      this._instance = new VerifyEmailService(
-        options.verifyEmailRepository,
-        options.prismaService,
-        options.userService,
-        options.jwtService,
-        options.loggerService
-      );
-    }
-    return this._instance;
-  }
 
   async createToken(
     tx: Prisma.TransactionClient,
@@ -92,37 +80,41 @@ export class VerifyEmailService implements IVerifyEmailService {
   }
 
   async verifyUserEmail(token: string): Promise<void> {
-    const payload = await this._jwtService.verify({
+    const { jwtPayload, error } = this._jwtService.verify({
       token,
       secret: VERIFY_EMAIL_TOKEN_SECRET,
     });
 
-    await this._prismaService.transaction(async (tx) => {
-      const foundToken = await this._verifyEmailRepository.findByToken(
-        token,
-        tx
-      );
-
-      if (!foundToken) {
-        this._loggerService.warn(
-          `Token reuse attempt for user ${payload.userId}`
+    if (jwtPayload) {
+      await this._prismaService.transaction(async (tx) => {
+        const foundToken = await this._verifyEmailRepository.findByToken(
+          token,
+          tx
         );
 
-        await this._verifyEmailRepository.deleteMany(payload.userId, tx);
-        throw new BadRequestError(ERROR_MESSAGES.INVALID_TOKEN);
-      }
+        if (!foundToken) {
+          this._loggerService.warn(
+            `Token reuse attempt for user ${jwtPayload.userId}`
+          );
 
-      // Deletes the email verification token from the database.
-      await this._verifyEmailRepository.deleteMany(foundToken.userId, tx);
+          await this._verifyEmailRepository.deleteMany(jwtPayload.userId, tx);
+          throw new BadRequestError(ERROR_MESSAGES.INVALID_TOKEN);
+        }
 
-      const verifiedUser = await this._userService.verifyUserEmail(
-        foundToken.userId,
-        tx
-      );
+        // Deletes the email verification token from the database.
+        await this._verifyEmailRepository.deleteMany(foundToken.userId, tx);
 
-      this._loggerService.info(
-        `Email verification successful for user ${verifiedUser.id}`
-      );
-    });
+        const verifiedUser = await this._userService.verifyUserEmail(
+          foundToken.userId,
+          tx
+        );
+
+        this._loggerService.info(
+          `Email verification successful for user ${verifiedUser.id}`
+        );
+      });
+    } else {
+      throw error;
+    }
   }
 }
